Prepare project for GitHub: ignore secrets, add env templates, update README

+++ src/Pages/Users/User.js
@@ -1,15 +1,15 @@
 import axios from "axios";
-import "../../style.css";
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router";
+import "../../styles/components.css";
 
-const User = ({ user,getUserUpdates }) => {
+const User = ({ user, getUserUpdates }) => {
   const { profiles: usersProfile } = useSelector((state) => state.profiles);
   const { permissions: usersPermissions } = useSelector(
     (state) => state.permissions
   );
-  const dispatch=useDispatch();
+  const dispatch = useDispatch();
   const [response, setResponse] = useState("");
   const navigate = useNavigate();
 
@@ -22,15 +22,15 @@ const User = ({ user,getUserUpdates }) => {
   )?.permissions||[];
   
   const deleteUser = async () => {
-    dispatch({type:"DELETE",payload:user._id,reducerKey:'users'})
-    dispatch({type:"DELETE",payload:user._id,reducerKey:'profiles'})
-    dispatch({type:"DELETE",payload:user._id,reducerKey:'permissions'})
+    if (window.confirm("Are you sure you want to delete this user?")) {
+      dispatch({type:"DELETE",payload:user._id,reducerKey:'users'})
+      dispatch({type:"DELETE",payload:user._id,reducerKey:'profiles'})
+      dispatch({type:"DELETE",payload:user._id,reducerKey:'permissions'})
 
-
-    const resp = await axios.delete(`http://localhost:4001/user/${user._id}`);
-    getUserUpdates(user._id);
-    setResponse(resp);
-    console.log(response);
+      const resp = await axios.delete(`http://localhost:4001/user/${user._id}`);
+      getUserUpdates(user._id);
+      setResponse(resp);
+    }
   };
 
   const editUser = () => {
@@ -42,19 +42,35 @@ const User = ({ user,getUserUpdates }) => {
   },[profile])
 
   return (
-    <div className="page-border2">
-      Name : {profile?.FirstName} {profile?.LastName} <br />
-      User Name : {user.user} <br />
-      Session time out {"(Minutes)"} : {profile?.SessionTimeOut} <br />
-      Created date : {profile?.CreatedDate}
-      <br />
-      Permissions:
-      {userPermissions.map((userP,index) => {
-        return <span key={index}> {userP} , </span>;
-      })}
-      <br />
-      <button style={{marginTop:"10px",marginBottom:"3px"}} onClick={() => editUser()}>Edit</button>
-      <button onClick={() => deleteUser()}>Delete</button>
+    <div className="user-card">
+      <div className="user-info">
+        <h3 className="user-name">{profile?.FirstName} {profile?.LastName}</h3>
+        <p className="user-detail"><strong>Username:</strong> {user.user}</p>
+        <p className="user-detail"><strong>Session Timeout:</strong> {profile?.SessionTimeOut} minutes</p>
+        <p className="user-detail"><strong>Created:</strong> {profile?.CreatedDate}</p>
+        <div className="user-permissions">
+          <strong>Permissions:</strong>
+          <div className="permissions-list">
+            {userPermissions.map((userP, index) => (
+              <span key={index} className="permission-tag">{userP}</span>
+            ))}
+          </div>
+        </div>
+      </div>
+      <div className="user-actions">
+        <button 
+          className="edit-button"
+          onClick={editUser}
+        >
+          Edit
+        </button>
+        <button 
+          className="delete-button"
+          onClick={deleteUser}
+        >
+          Delete
+        </button>
+      </div>
     </div>
   );
 };
